backend/agents/robust_manim_agent/requirements.txt
google-adk
manim
python-dotenv
vertexai>=1.0.0 